+++ src/compose.js
@@ -11,6 +11,10 @@
  * @symb R.compose(f, g, h)(a, b) = f(g(h(a, b)))
  */
 export function compose(...fns){
+  if(fns.length === 0){
+    throw new Error('compose requires at least one argument')
+  } 
+
   return (...args) => {
     const list = fns.slice()
     if (list.length > 0){
@@ -22,7 +26,5 @@ export function compose(...fns){
 
       return result
     }
-
-    return undefined
   }
 }