+++ src/drop.js
@@ -1,5 +1,5 @@
 /**
- * Returns all but the first `n` elements of the given list, string, or
+ * Returns all but the first `n` elements of the given listOrString, string, or
  * transducer/transformer (or object with a `drop` method).
  *
  * @func
@@ -8,7 +8,7 @@
  * @sig Number -> String -> String
  * @param {Number} n
  * @param {*} list
- * @return {*} A copy of list without the first `n` elements
+ * @return {*} A copy of listOrString without the first `n` elements
  * @example
  *
  *      R.drop(1, ['foo', 'bar', 'baz']); //=> ['bar', 'baz']
@@ -17,8 +17,8 @@
  *      R.drop(4, ['foo', 'bar', 'baz']); //=> []
  *      R.drop(3, 'ramda');               //=> 'da'
  */
-export function drop(n, list){
+export function drop(n, listOrString){
   if (arguments.length === 1) return _list => drop(n, _list)
-
-  return list.slice(n)
+  if(n< 0) return listOrString
+  return listOrString.slice(n)
 }