+++ src/ifElse.spec.js
@@ -1,17 +1,36 @@
 import { has } from './has'
+import { identity } from './identity.js'
 import { prop } from './prop'
 import { always } from './always'
 import { ifElse } from './ifElse'
 
 const condition = has('foo')
+const v = function(a){ return typeof a === 'number' }
+const t = function(a){ return a + 1 }
 const ifFn = x => prop('foo', x).length
 const elseFn = () => false
 
-test('', () => {
+test('happy', () => {
   const fn = ifElse(condition, ifFn)(elseFn)
 
   expect(fn({ foo : 'bar' })).toEqual(3)
   expect(fn({ fo : 'bar' })).toEqual(false)
+})
+
+test('ramda spec', () => {
+  const ifIsNumber = ifElse(v)
+  expect(ifIsNumber(t, identity)(15)).toEqual(16)
+  expect(ifIsNumber(t, identity)('hello')).toEqual('hello')
+})
+
+test('pass all arguments', () => {
+  const identity = function(a){ return a }
+  const v = function(){ return true }
+  const onTrue = function(a, b){
+    expect(a).toEqual(123)
+    expect(b).toEqual('abc')
+  }
+  ifElse(v, onTrue, identity)(123, 'abc')
 })
 
 test('accept constant as condition', () => {
@@ -26,14 +45,14 @@ test('accept constant as condition - case 2', () => {
   expect(fn()).toEqual(false)
 })
 
-test('curry (x)(y,z)', () => {
+test('curry 1', () => {
   const fn = ifElse(condition, ifFn)(elseFn)
 
   expect(fn({ foo : 'bar' })).toEqual(3)
   expect(fn({ fo : 'bar' })).toEqual(false)
 })
 
-test('curry (x)(y)(z)', () => {
+test('curry 2', () => {
   const fn = ifElse(condition)(ifFn)(elseFn)
 
   expect(fn({ foo : 'bar' })).toEqual(3)