+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 - 4.0.0 Multiple breaking changes as Rambda methods are changed in order to increase the similarity between with Ramda
 
+All breaking changes:
+
+-- R.add works only with numbers
+
+-- Fix R.adjust which had wrong order of arguments
+
 - 3.3.0 Close [issue #245](https://github.com/selfrefactor/rambda/issues/245) - complete typings tests for methods that have more specific Typescript definitions
 
 - 3.2.1 Fast fix for [issue #273](https://github.com/selfrefactor/rambda/issues/273) - messed up typings