+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All breaking changes:
 
 -- Fix R.adjust which had wrong order of arguments
 
+-- R.adjust works when index is out of bounds
+
 - 3.3.0 Close [issue #245](https://github.com/selfrefactor/rambda/issues/245) - complete typings tests for methods that have more specific Typescript definitions
 
 - 3.2.1 Fast fix for [issue #273](https://github.com/selfrefactor/rambda/issues/273) - messed up typings