+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All breaking changes:
 
 -- R.adjust works when index is out of bounds
 
+-- R.complement support function with multiple arguments
+
 - 3.3.0 Close [issue #245](https://github.com/selfrefactor/rambda/issues/245) - complete typings tests for methods that have more specific Typescript definitions
 
 - 3.2.1 Fast fix for [issue #273](https://github.com/selfrefactor/rambda/issues/273) - messed up typings