+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Add to `Differences`:
 
 ```text
+R.type can return 'NaN'
+
 R.compose doesn't pass `this` context
 
 R.clone doesn't work with number, booleans and strings as input
@@ -23,6 +25,10 @@ All breaking changes:
 -- R.clone works with `Date` value as input
 
 -- R.drop/dropLast always return new copy of the list/string
+
+-- R.equals handles `NaN`
+
+-- R.type supports `new Boolean` expressions
 
 - 3.3.0 Close [issue #245](https://github.com/selfrefactor/rambda/issues/245) - complete typings tests for methods that have more specific Typescript definitions